misinfo-ai/requirements.txt
beautifulsoup4
requests
flask
flask-cors
python-dotenv
groq
